requirements.txt
aiogram==3.15.0
python-docx
Pillow
fpdf==1.7.2
sqlalchemy